requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
Django==3.2
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==3.10.1
kombu==5.0.2
pkg-resources==0.0.0
plaid-python==7.2.1
prompt-toolkit==3.0.18
pycparser==2.20
pytz==2021.1
redis==3.5.3
requests==2.25.1
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
